Update style.css
the buttons are at the bottom of the list
this means when you have a lot of character slots you have to scroll to the bottom to get to the button.

I made it so you don't have to scroll to see the buttons by moving the overflow and scroll buttons from the characters-list div to the characters div and I resized the characters element to allow the buttons to be viewable.