Update main.lua
Fixed issue #231 Getting stuck at black screen after selecting character and changing Config.SkipSelection to true. The issue seems to be the function incapsulates all spawning conditions while relying on a result from the OwnedApartment callback. This results in getting stuck at black screen when the player does not own an apartment and is attempting to spawn at last location. 
Also, updated RegisterNuiCallback('closeUI') on line 169 to use argument data as the previous method of using _ resulted in a declaration error on line 170 when attmepting to initialze cData = data.cData